+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 I will record changes to this file just so i don't need to look at commit history every time
 
 # Changelog
+2025-12-21:
+- Added script for updating mac on schedule
 2025-11-09:
 - Updated the script for updating linux on schedule to fix a bug where it was running before network was up and homebrew update was failing.